requirements.txt
pvporcupine
pyaudio
speechrecognition
pyttsx3
requests
python-dotenv
